ss-agents/pyproject.toml
[tool.poetry]
name = "ss-agents"
version = "0.1.0"
description = "This is the rig for stdio mcps with langgraph orchestration"
authors = [ "The Moose" ]
readme = "README.md"

# Tells Poetry these folders are "installable" packages
packages = [
  { include = "agents" },
  { include = "langgraph" },
  { include = "utils" }
]

# These are your main (runtime) dependencies
[tool.poetry.dependencies]
python = ">=3.12,<4.0"
langchain = "*"
langgraph = "*"
"langchain-mcp-adapters" = "*"
"python-dotenv" = "*"
requests = "*"
"langchain-openai" = ">=0.3.11,<0.4.0"

# dev dependencies go here (like Pytest)
[tool.poetry.group.dev.dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
